🔧 Refactor: Modular code structure + Security improvements
Major changes:
✅ Split const.py into modular structure (1,180 → 282 lines, -76%)
  - const_api.py: API endpoints, actions, rate limiting
  - const_devices.py: Device parameters, state mappings
  - const_sensors.py: Sensor definitions, units
  - const_features.py: Features, switches, setpoints

✅ Add API Rate Limiter (utils_rate_limiter.py)
  - Token bucket algorithm for burst protection
  - Configurable request limits (10 req/s, burst: 3)
  - Priority queue for critical requests
  - Statistics & monitoring

✅ Add Input Sanitization (utils_sanitizer.py)
  - XSS protection (HTML escaping)
  - Input validation for all data types
  - Range validation (temp, pH, ORP, chlorine)
  - Device key & API parameter validation

Benefits:
- Better code maintainability (modular structure)
- Improved security (input sanitization)
- API protection (rate limiting)
- Easier testing (smaller modules)

No breaking changes - all imports remain compatible.
